Renamed "check_data_type.py" to "decorators.py"
